Chore: #17 타이머매니저 파일 분리
